+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Update checker is collapsed by default.
 - Projects are collapsed when click on Projects.
 - Improve perfomance during adding new image.
+- Screenshot preview screen became of same size as screen.
 
 ## Bug Fixes
 - After making trims, if i select all area and copy - initial image is copied.